fix: introspect command
