+++ cli/src/commands/subgraph/commands/introspect.ts
@@ -1,46 +1,40 @@
-import { EnumStatusCode } from '@wundergraph/cosmo-connect/dist/common_pb';
+import { splitLabel } from '@wundergraph/cosmo-shared';
 import { Command } from 'commander';
 import pc from 'picocolors';
-import { splitLabel } from '@wundergraph/cosmo-shared';
-import { baseHeaders } from '../../../core/config.js';
 import { BaseCommandOptions } from '../../../core/types/types.js';
 import program from '../../index.js';
+import { introspectSubgraph } from '../utils.js';
 
 export default (opts: BaseCommandOptions) => {
-  const introspectSubgraph = new Command('introspect');
-  introspectSubgraph.description('Introspects a subgraph.');
-  introspectSubgraph.requiredOption(
+  const introspectSubgraphCmd = new Command('introspect');
+  introspectSubgraphCmd.description('Introspects a subgraph.');
+  introspectSubgraphCmd.requiredOption(
     '-r, --routing-url <url>',
     'The routing url of your subgraph. This is the url that the subgraph will be accessible at.',
   );
-  introspectSubgraph.option(
+  introspectSubgraphCmd.option(
     '--header [headers...]',
     'The headers to apply when the subgraph is introspected. This is used for authentication and authorization.The headers are passed in the format <key>=<value> <key>=<value>.Use quotes if there exists space in the key/value.',
   );
-  introspectSubgraph.action(async (options) => {
-    const resp = await opts.client.platform.introspectSubgraph(
-      {
-        url: options.routingUrl,
-        headers: options.header?.map((label: string) => {
+  introspectSubgraphCmd.action(async (options) => {
+    const resp = await introspectSubgraph({
+      subgraphURL: options.routingUrl,
+      additionalHeaders:
+        options.header?.map((label: string) => {
           const { key, value } = splitLabel(label);
           return {
             key,
             value,
           };
-        }),
-      },
-      {
-        headers: baseHeaders,
-      },
-    );
+        }) || [],
+    });
 
-    if (resp.response?.code === EnumStatusCode.OK) {
-      console.log('SDL of the subgraph:-\n');
+    if (resp.success === true) {
       console.log(resp.sdl);
     } else {
       program.error(pc.red('Could not introspect the subgraph.'));
     }
   });
 
-  return introspectSubgraph;
+  return introspectSubgraphCmd;
 };
