+++ connect/src/wg/cosmo/platform/v1/platform_pb.ts
@@ -4447,89 +4447,3 @@ export class Header extends Message<Header> {
   }
 }
 
-/**
- * @generated from message wg.cosmo.platform.v1.IntrospectSubgraphRequest
- */
-export class IntrospectSubgraphRequest extends Message<IntrospectSubgraphRequest> {
-  /**
-   * @generated from field: string url = 1;
-   */
-  url = "";
-
-  /**
-   * @generated from field: repeated wg.cosmo.platform.v1.Header headers = 2;
-   */
-  headers: Header[] = [];
-
-  constructor(data?: PartialMessage<IntrospectSubgraphRequest>) {
-    super();
-    proto3.util.initPartial(data, this);
-  }
-
-  static readonly runtime: typeof proto3 = proto3;
-  static readonly typeName = "wg.cosmo.platform.v1.IntrospectSubgraphRequest";
-  static readonly fields: FieldList = proto3.util.newFieldList(() => [
-    { no: 1, name: "url", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 2, name: "headers", kind: "message", T: Header, repeated: true },
-  ]);
-
-  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): IntrospectSubgraphRequest {
-    return new IntrospectSubgraphRequest().fromBinary(bytes, options);
-  }
-
-  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): IntrospectSubgraphRequest {
-    return new IntrospectSubgraphRequest().fromJson(jsonValue, options);
-  }
-
-  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): IntrospectSubgraphRequest {
-    return new IntrospectSubgraphRequest().fromJsonString(jsonString, options);
-  }
-
-  static equals(a: IntrospectSubgraphRequest | PlainMessage<IntrospectSubgraphRequest> | undefined, b: IntrospectSubgraphRequest | PlainMessage<IntrospectSubgraphRequest> | undefined): boolean {
-    return proto3.util.equals(IntrospectSubgraphRequest, a, b);
-  }
-}
-
-/**
- * @generated from message wg.cosmo.platform.v1.IntrospectSubgraphResponse
- */
-export class IntrospectSubgraphResponse extends Message<IntrospectSubgraphResponse> {
-  /**
-   * @generated from field: wg.cosmo.platform.v1.Response response = 1;
-   */
-  response?: Response;
-
-  /**
-   * @generated from field: string sdl = 2;
-   */
-  sdl = "";
-
-  constructor(data?: PartialMessage<IntrospectSubgraphResponse>) {
-    super();
-    proto3.util.initPartial(data, this);
-  }
-
-  static readonly runtime: typeof proto3 = proto3;
-  static readonly typeName = "wg.cosmo.platform.v1.IntrospectSubgraphResponse";
-  static readonly fields: FieldList = proto3.util.newFieldList(() => [
-    { no: 1, name: "response", kind: "message", T: Response },
-    { no: 2, name: "sdl", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-  ]);
-
-  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): IntrospectSubgraphResponse {
-    return new IntrospectSubgraphResponse().fromBinary(bytes, options);
-  }
-
-  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): IntrospectSubgraphResponse {
-    return new IntrospectSubgraphResponse().fromJson(jsonValue, options);
-  }
-
-  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): IntrospectSubgraphResponse {
-    return new IntrospectSubgraphResponse().fromJsonString(jsonString, options);
-  }
-
-  static equals(a: IntrospectSubgraphResponse | PlainMessage<IntrospectSubgraphResponse> | undefined, b: IntrospectSubgraphResponse | PlainMessage<IntrospectSubgraphResponse> | undefined): boolean {
-    return proto3.util.equals(IntrospectSubgraphResponse, a, b);
-  }
-}
-
