--- controlplane/src/core/services/IntrospectSubgraph.ts
@@ -1,49 +1,0 @@
-import { Header } from 'src/types/index.js';
-
-export default class IntrospectSubgraph {
-  constructor() {}
-
-  public async introspect({
-    subgraphURL,
-    additionalHeaders,
-  }: {
-    subgraphURL: string;
-    additionalHeaders: Header[];
-  }): Promise<{ sdl: string; errorMessage?: string; success: boolean }> {
-    const headers = new Headers();
-    headers.append('Content-Type', 'application/json');
-    for (const header of additionalHeaders) {
-      headers.append(header.key, header.value);
-    }
-
-    const graphql = JSON.stringify({
-      query: `
-        {
-          _service{
-            sdl
-          }
-        }
-      `,
-      variables: {},
-    });
-
-    const response = await fetch(subgraphURL, {
-      method: 'POST',
-      headers,
-      body: graphql,
-    });
-    if (response.status !== 200) {
-      return {
-        success: false,
-        errorMessage: 'Could not introspect the subgraph.',
-        sdl: '',
-      };
-    }
-    const body = await response.json();
-    const data = body.data;
-    return {
-      success: true,
-      sdl: data._service.sdl,
-    };
-  }
-}
